requirements.txt
pandas
numpy
scikit-learn==1.7.2
joblib
flask
pyyaml
xgboost
catboost